Kendall-package {Kendall}	R Documentation

Kendall correlation and trend tests.

Description

Compute Kendall's rank correlation and various trend tests.

Details

Package: 	 Kendall

Type: 	 Package

Version: 	 2.2

Date: 	 2011-05-16

License: 	 GPL (>= 2)

LazyLoad: 	 yes

If you just want to compute Kendall's tau or its signficance level, the base function cor and cor.test are recommended. The purpose of this package is to implement the Mann-Kendall test, the seasonal Mann-Kendall trend test as well as computing the Kendall score. See documentation for Mann-Kendall on using the block bootstrap to deal with autocorrelation.

Author(s)

A. I. McLeod

Maintainer: aimcleod@uwo.ca

References

Hipel, K.W. and McLeod, A.I., (1994). Time Series Modelling of Water Resources and Environmental Systems.

See Also

cor, cor.test, print.Kendall, summary.Kendall, MannKendall, SeasonalMannKendall

Examples


library(boot)
data(PrecipGL)
MKtau<-function(z) MannKendall(z)$tau
tsboot(PrecipGL, MKtau, R=500, l=5, sim="fixed")

[Package Kendall version 2.2.1 Index]