print.Kendall {Kendall}	R Documentation

print Method for Class 'Kendall'

Description

The value of Kendall's tau and its two-sided p-value are displayed.

Usage


## S3 method for class 'Kendall'
print(x, ...)

Arguments

x
output from Kendall, MannKendall or SeasonalMannKendall

...
any additional arguments

Value

NULL

Note

If you want to use the output from Kendall or the other functions, save the result as in res<-Kendall(x,y) and then select from res the value(s) needed.

Author(s)

A.I. McLeod, aimcleod@uwo.ca

See Also

summary.Kendall, Kendall, MannKendall,SeasonalMannKendall, summary.Kendall

Examples


x<-1:10
y<-rnorm(10)
Kendall(x,y)
#is equivalent to
out<-Kendall(x,y)
print(out)

[Package Kendall version 2.2.1 Index]